backend/aegisalert/users/serializers.py
from rest_framework import serializers
from .models import User
from .models import Notification

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'password', 'email', 'phone_number', 'address', 'first_name', 'last_name']
        extra_kwargs = {
            'password': {'write_only': True},  # Ensure password is write-only
        }

    def create(self, validated_data):
        # Use Django's built-in `create_user` method to handle password hashing
        user = User.objects.create_user(**validated_data)
        return user
    
    
class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = ['id', 'message', 'created_at', 'is_read']